util/codec: remove old bytes codec.
